+++ R/random_mctq.R
@@ -436,7 +436,7 @@ random_shift_mctq <- function(n_w = c(n_w_m = 6, n_w_e = 4, n_w_n = 6),
     for (i in c("bt_w_m", "bt_w_e", "bt_w_n", "bt_f_m", "bt_f_e", "bt_f_n")) {
         sample <- sample_time(min = min, max = max, by = by, prob = prob)
         name <- paste0("sprep_", str_extract_(i, "._.$"))
-        assign(i, sum_time(get(name), - sample, circular = TRUE))
+        assign(i, sum_time(get(name), - sample, cycle = lubridate::ddays()))
     }
 
     # Create `slat_w_*` and `slat_f_*` -----
@@ -510,7 +510,7 @@ random_shift_mctq <- function(n_w = c(n_w_m = 6, n_w_e = 4, n_w_n = 6),
         se_w_m = list(
             name = "se_w_m",
             min = sum_time(sprep_w_m, slat_w_m, hms::parse_hm("02:05"),
-                           circular = TRUE),
+                           cycle = lubridate::ddays()),
             max = hms::parse_hm("05:00"), # Changed
             mean = hms::parse_hm("04:35"),
             sd = hms::parse_hm("00:35")),
@@ -521,7 +521,7 @@ random_shift_mctq <- function(n_w = c(n_w_m = 6, n_w_e = 4, n_w_n = 6),
         se_w_e = list(
             name = "se_w_e",
             min = sum_time(sprep_w_e, slat_w_e, hms::parse_hm("03:40"),
-                           circular = TRUE),
+                           cycle = lubridate::ddays()),
             max = hms::parse_hm("12:25"),
             mean = hms::parse_hm("08:25"),
             sd = hms::parse_hm("01:20")),
@@ -532,7 +532,7 @@ random_shift_mctq <- function(n_w = c(n_w_m = 6, n_w_e = 4, n_w_n = 6),
         se_w_n = list(
             name = "se_w_n",
             min = sum_time(sprep_w_n, slat_w_n, hms::parse_hm("02:00"), # Chan.
-                           circular = TRUE),
+                           cycle = lubridate::ddays()),
             max = hms::parse_hm("18:05"),
             mean = hms::parse_hm("13:30"),
             sd = hms::parse_hm("01:30")),
@@ -678,7 +678,7 @@ random_shift_mctq <- function(n_w = c(n_w_m = 6, n_w_e = 4, n_w_n = 6),
                 "nape_f_n")) {
         name <- get(paste0("napo_", str_extract_(i, "._.$")))
         sample <- sample_time(min = min, max = max, by = by, prob = prob)
-        assign(i, sum_time(name, sample, circular = TRUE))
+        assign(i, sum_time(name, sample, cycle = lubridate::ddays()))
     }
 
     # Create `nap_w_*` and `nap_f_*` -----