+++ R/tbt.R
@@ -96,5 +96,6 @@ tbt <- function(bt, gu) {
     checkmate::assert_class(gu, "hms")
     assert_identical(bt, gu, type = "length")
 
-    sum_time(gu, - bt, class = "Duration", circular = TRUE, vectorize = TRUE)
+    sum_times(gu, - bt, cycle = lubridate::ddays()) %>%
+        lubridate::as.duration()
 }