+++ tests/testthat/test-sum_time.R
@@ -1,4 +1,4 @@
-test_that("sum_time() | non-vectorized test", {
+test_that("sum_time() | linear test", {
     expect_equal(sum_time(c(lubridate::dhours(1), lubridate::dminutes(30)),
                           lubridate::hours(1),
                           c(as.difftime(1, units = "hours"),
@@ -9,11 +9,40 @@ test_that("sum_time() | non-vectorized test", {
                               "1970-01-01 01:00:00")),
                           lubridate::as.interval(lubridate::dhours(1),
                                                  as.Date("1970-01-01")),
-                          class = "duration",
-                          circular = FALSE,
-                          vectorize = FALSE,
-                          na.rm = FALSE),
-                 lubridate::dhours(30)) # 30:00:00
+                          cycle = NULL,
+                          na_rm = FALSE),
+                 hms::hms(as.numeric(lubridate::dhours(30)))) # 30:00:00
+    expect_equal(sum_time(hms::as_hms(NA), # !
+                          c(lubridate::dhours(1), lubridate::dminutes(30)),
+                          lubridate::hours(1),
+                          c(as.difftime(1, units = "hours"),
+                            as.difftime(30, units = "mins")),
+                          c(hms::parse_hm("02:00"), hms::parse_hm("02:00")),
+                          lubridate::as_datetime("1970-01-01 20:00:00"),
+                          as.POSIXlt(lubridate::as_datetime(
+                              "1970-01-01 01:00:00")),
+                          lubridate::as.interval(lubridate::dhours(1),
+                                                 as.Date("1970-01-01")),
+                          cycle = NULL,
+                          na_rm = FALSE), # !
+                 hms::as_hms(NA))
+    expect_equal(sum_time(hms::as_hms(NA), # !
+                          c(lubridate::dhours(1), lubridate::dminutes(30)),
+                          lubridate::hours(1),
+                          c(as.difftime(1, units = "hours"),
+                            as.difftime(30, units = "mins")),
+                          c(hms::parse_hm("02:00"), hms::parse_hm("02:00")),
+                          lubridate::as_datetime("1970-01-01 20:00:00"),
+                          as.POSIXlt(lubridate::as_datetime(
+                              "1970-01-01 01:00:00")),
+                          lubridate::as.interval(lubridate::dhours(1),
+                                                 as.Date("1970-01-01")),
+                          cycle = NULL,
+                          na_rm = TRUE), # !
+                 hms::hms(as.numeric(lubridate::dhours(30)))) # 30:00:00
+})
+
+test_that("sum_time() | circular test", {
     expect_equal(sum_time(c(lubridate::dhours(1), lubridate::dminutes(30)),
                           lubridate::hours(1),
                           c(as.difftime(1, units = "hours"),
@@ -24,43 +53,9 @@ test_that("sum_time() | non-vectorized test", {
                               "1970-01-01 01:00:00")),
                           lubridate::as.interval(lubridate::dhours(1),
                                                  as.Date("1970-01-01")),
-                          class = "hms",
-                          circular = TRUE, # !
-                          vectorize = FALSE,
-                          na.rm = FALSE),
+                          cycle = lubridate::ddays(), # !
+                          na_rm = FALSE),
                  hms::parse_hm("06:00")) # 06:00 | 30 - 24
-    expect_equal(sum_time(hms::as_hms(NA), # !
-                          c(lubridate::dhours(1), lubridate::dminutes(30)),
-                          lubridate::hours(1),
-                          c(as.difftime(1, units = "hours"),
-                            as.difftime(30, units = "mins")),
-                          c(hms::parse_hm("02:00"), hms::parse_hm("02:00")),
-                          lubridate::as_datetime("1970-01-01 20:00:00"),
-                          as.POSIXlt(lubridate::as_datetime(
-                              "1970-01-01 01:00:00")),
-                          lubridate::as.interval(lubridate::dhours(1),
-                                                 as.Date("1970-01-01")),
-                          class = "period",
-                          circular = FALSE,
-                          vectorize = FALSE,
-                          na.rm = FALSE), # !
-                 lubridate::as.period(NA))
-    expect_equal(sum_time(hms::as_hms(NA), # !
-                          c(lubridate::dhours(1), lubridate::dminutes(30)),
-                          lubridate::hours(1),
-                          c(as.difftime(1, units = "hours"),
-                            as.difftime(30, units = "mins")),
-                          c(hms::parse_hm("02:00"), hms::parse_hm("02:00")),
-                          lubridate::as_datetime("1970-01-01 20:00:00"),
-                          as.POSIXlt(lubridate::as_datetime(
-                              "1970-01-01 01:00:00")),
-                          lubridate::as.interval(lubridate::dhours(1),
-                                                 as.Date("1970-01-01")),
-                          class = "period",
-                          circular = FALSE,
-                          vectorize = FALSE,
-                          na.rm = TRUE), # !
-                 lubridate::as.period(hms::hms(108000))) # 30:00:00
     expect_equal(sum_time(hms::as_hms(NA),
                           c(lubridate::dhours(1), lubridate::dminutes(30)),
                           lubridate::hours(1),
@@ -72,137 +67,25 @@ test_that("sum_time() | non-vectorized test", {
                               "1970-01-01 01:00:00")),
                           lubridate::as.interval(lubridate::dhours(1),
                                                  as.Date("1970-01-01")),
-                          class = "difftime",
-                          circular = TRUE, # !
-                          vectorize = FALSE,
-                          na.rm = TRUE),
-                 lubridate::as.difftime(21600, units = "secs"))
+                          cycle = lubridate::ddays(),
+                          na_rm = TRUE),
+                 hms::parse_hm("06:00"))
 })
 
-test_that("sum_time()| vectorized test", {
-    expect_equal(sum_time(c(lubridate::dhours(1), lubridate::dminutes(30)),
-                          c(lubridate::hours(1), lubridate::hours(1)),
-                          c(as.difftime(1, units = "hours"),
-                            as.difftime(30, units = "mins")),
-                          c(hms::parse_hm("02:00"), hms::parse_hm("02:00")),
-                          c(lubridate::as_datetime("1970-01-01 20:00:00"),
-                            lubridate::as_datetime("1970-01-01 10:00:00")),
-                          as.POSIXlt(
-                              c(lubridate::as_datetime("1970-01-01 01:00:00"),
-                                lubridate::as_datetime("1970-01-01 02:00:00"))),
-                          c(lubridate::as.interval(lubridate::dhours(4),
-                                                   as.Date("1970-01-01")),
-                            lubridate::as.interval(lubridate::dhours(1),
-                                                   as.Date("1970-01-01"))),
-                          class = "duration",
-                          circular = FALSE,
-                          vectorize = TRUE,
-                          na.rm = FALSE),
-                 c(lubridate::dhours(30),
-                   lubridate::dhours(17))) # 30:00:00 | 17:00:00
-    expect_equal(sum_time(c(lubridate::dhours(1), lubridate::dminutes(30)),
-                          c(lubridate::hours(1), lubridate::hours(1)),
-                          c(as.difftime(1, units = "hours"),
-                            as.difftime(30, units = "mins")),
-                          c(hms::parse_hm("02:00"), hms::parse_hm("02:00")),
-                          c(lubridate::as_datetime("1970-01-01 20:00:00"),
-                            lubridate::as_datetime("1970-01-01 10:00:00")),
-                          as.POSIXlt(
-                              c(lubridate::as_datetime("1970-01-01 01:00:00"),
-                                lubridate::as_datetime("1970-01-01 02:00:00"))),
-                          c(lubridate::as.interval(lubridate::dhours(4),
-                                                   as.Date("1970-01-01")),
-                            lubridate::as.interval(lubridate::dhours(1),
-                                                   as.Date("1970-01-01"))),
-                          class = "hms",
-                          circular = TRUE, # !
-                          vectorize = TRUE,
-                          na.rm = FALSE),
-                 c(hms::parse_hm("06:00"),
-                   hms::parse_hm("17:00"))) # 06:00:00 | 17:00:00
-    expect_equal(sum_time(c(hms::as_hms(NA), hms::as_hms(NA)), # !
-                          c(lubridate::dhours(1), lubridate::dminutes(30)),
-                          c(lubridate::hours(1), lubridate::hours(1)),
-                          c(as.difftime(1, units = "hours"),
-                            as.difftime(30, units = "mins")),
-                          c(hms::parse_hm("02:00"), hms::parse_hm("02:00")),
-                          c(lubridate::as_datetime("1970-01-01 20:00:00"),
-                            lubridate::as_datetime("1970-01-01 10:00:00")),
-                          as.POSIXlt(
-                              c(lubridate::as_datetime("1970-01-01 01:00:00"),
-                                lubridate::as_datetime("1970-01-01 02:00:00"))),
-                          c(lubridate::as.interval(lubridate::dhours(4),
-                                                   as.Date("1970-01-01")),
-                            lubridate::as.interval(lubridate::dhours(1),
-                                                   as.Date("1970-01-01"))),
-                          class = "period",
-                          circular = FALSE,
-                          vectorize = TRUE,
-                          na.rm = FALSE), # !
-                 c(lubridate::as.period(NA),
-                   lubridate::as.period(NA)))
-    expect_equal(sum_time(c(hms::as_hms(NA), hms::as_hms(NA)), # !
-                          c(lubridate::dhours(1), lubridate::dminutes(30)),
-                          c(lubridate::hours(1), lubridate::hours(1)),
-                          c(as.difftime(1, units = "hours"),
-                            as.difftime(30, units = "mins")),
-                          c(hms::parse_hm("02:00"), hms::parse_hm("02:00")),
-                          c(lubridate::as_datetime("1970-01-01 20:00:00"),
-                            lubridate::as_datetime("1970-01-01 10:00:00")),
-                          as.POSIXlt(
-                              c(lubridate::as_datetime("1970-01-01 01:00:00"),
-                                lubridate::as_datetime("1970-01-01 02:00:00"))),
-                          c(lubridate::as.interval(lubridate::dhours(4),
-                                                   as.Date("1970-01-01")),
-                            lubridate::as.interval(lubridate::dhours(1),
-                                                   as.Date("1970-01-01"))),
-                          class = "period",
-                          circular = FALSE,
-                          vectorize = TRUE,
-                          na.rm = TRUE), # !
-                 c(lubridate::as.period(hms::hms(108000)),
-                   lubridate::as.period(hms::hms(61200)))) # 30:00:00 | 17:00:00
-    expect_equal(sum_time(c(hms::as_hms(NA), hms::as_hms(NA)),
-                          c(lubridate::dhours(1), lubridate::dminutes(30)),
-                          c(lubridate::hours(1), lubridate::hours(1)),
-                          c(as.difftime(1, units = "hours"),
-                            as.difftime(30, units = "mins")),
-                          c(hms::parse_hm("02:00"), hms::parse_hm("02:00")),
-                          c(lubridate::as_datetime("1970-01-01 20:00:00"),
-                            lubridate::as_datetime("1970-01-01 10:00:00")),
-                          as.POSIXlt(
-                              c(lubridate::as_datetime("1970-01-01 01:00:00"),
-                                lubridate::as_datetime("1970-01-01 02:00:00"))),
-                          c(lubridate::as.interval(lubridate::dhours(4),
-                                                   as.Date("1970-01-01")),
-                            lubridate::as.interval(lubridate::dhours(1),
-                                                   as.Date("1970-01-01"))),
-                          class = "difftime",
-                          circular = TRUE, # !
-                          vectorize = TRUE,
-                          na.rm = TRUE),
-                 c(lubridate::as.difftime(21600, units = "secs"),
-                   lubridate::as.difftime(61200, units = "secs")))
-})
-
-test_that("sum_time() | error test", {
-    expect_error(sum_time(1, class = "hms", circular = TRUE,
-                          vectorize = TRUE, na.rm = TRUE),
+test_that("build_sum() | error test", {
+    expect_error(build_sum(1, vectorize = FALSE, cycle = NULL, na_rm = FALSE),
                  "Assertion on 'x' failed")
-    expect_error(sum_time(hms::hms(1), class = 1, circular = TRUE,
-                          vectorize = TRUE, na.rm = TRUE),
-                 "Assertion on 'tolower\\(class\\)' failed")
-    expect_error(sum_time(hms::hms(1), class = "hms", circular = "",
-                          vectorize = TRUE, na.rm = TRUE),
-                 "Assertion on 'circular' failed")
-    expect_error(sum_time(hms::hms(1), class = "hms", circular = TRUE,
-                          vectorize = "", na.rm = TRUE),
+    expect_error(build_sum(hms::hms(1), vectorize = "", cycle = NULL,
+                           na_rm = TRUE),
                  "Assertion on 'vectorize' failed")
-    expect_error(sum_time(hms::hms(1), class = "hms", circular = TRUE,
-                          vectorize = TRUE, na.rm = ""),
-                 "Assertion on 'na.rm' failed")
+    expect_error(build_sum(hms::hms(1), vectorize = FALSE, cycle = "",
+                           na_rm = FALSE),
+                 "Assertion on 'cycle' failed")
+    expect_error(build_sum(hms::hms(1), vectorize = FALSE, cycle = NULL,
+                           na_rm = ""),
+                 "Assertion on 'na_rm' failed")
 
-    expect_error(sum_time(hms::hms(1), c(hms::hms(1), hms::hms(1)),
+    expect_error(build_sum(hms::hms(1), c(hms::hms(1), hms::hms(1)),
                           vectorize = TRUE),
-                 "When 'vectorize' is 'TRUE', all values in '...' must ")
+                 "All values in '...' must ")
 })